feat(theme-toggle): update to use firestore
